+++ backend/app.js
@@ -15,37 +15,11 @@ const { PORT = 3000 } = process.env;
 
 const handleErrors = require('./middlewares/handleErrors');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-
-const allowedCors = [
-  'http://mesto.khaera.nomoredomains.xyz',
-  'https://mesto.khaera.nomoredomains.xyz',
-  'http://localhost:3000',
-  'https://localhost:3000',
-  'localhost:3000',
-];
+const cors = require('./middlewares/cors');
 
 const app = express();
 
-// eslint-disable-next-line consistent-return
-app.use((req, res, next) => {
-  const { origin } = req.headers;
-  const { method } = req;
-  const requestHeaders = req.headers['access-control-request-headers'];
-  const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
-
-  if (allowedCors.includes(origin)) {
-    res.header('Access-Control-Allow-Origin', origin);
-  }
-
-  if (method === 'OPTIONS') {
-    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
-    res.header('Access-Control-Allow-Headers', requestHeaders);
-    return res.end();
-  }
-
-  next();
-});
-
+app.use(cors);
 app.use(bodyParser.json());
 
 mongoose.connect('mongodb://localhost:27017/mestodb', {
