+++ backend/middlewares/cors.js
@@ -1,0 +1,27 @@
+const allowedCors = [
+  'http://mesto.khaera.nomoredomains.xyz',
+  'https://mesto.khaera.nomoredomains.xyz',
+  'http://localhost:3000',
+  'https://localhost:3000',
+  'localhost:3000',
+];
+
+// eslint-disable-next-line consistent-return
+module.exports = (req, res, next) => {
+  const { origin } = req.headers;
+  const { method } = req;
+  const requestHeaders = req.headers['access-control-request-headers'];
+  const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
+
+  if (allowedCors.includes(origin)) {
+    res.header('Access-Control-Allow-Origin', origin);
+  }
+
+  if (method === 'OPTIONS') {
+    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
+    res.header('Access-Control-Allow-Headers', requestHeaders);
+    return res.end();
+  }
+
+  next();
+};
